AUTOPILOT_SYSTEM/backend/requirements.txt
# Certification Autopilot System - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
supabase==2.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# PDF Processing
pypdf==4.0.1
reportlab==4.0.9
pillow==10.2.0

# Browser Automation
playwright==1.41.0

# HTTP & API
httpx==0.26.0
requests==2.31.0

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2

# File Storage
boto3==1.34.34
minio==7.2.3

# Task Queue (Optional)
celery==5.3.6
redis==5.0.1

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.40.0

# Email
python-smtp==1.0.0
sendgrid==6.11.0

# Utilities
python-slugify==8.0.1
pytz==2024.1
